// src/utils/style/style.type.ts
export type Colors = {
  primary: string;    // Main background color
  secondary: string;  // Text and icons
  border: string;     // Border and divider lines
  accent: string;     // Accent color for buttons and links
  error: string;      // Error states
  success: string;    // Success states
  warning: string;    // Warning states
  info: string;       // Info states
  lightPrimary: string
  lightSecondary: string
};